requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloud-sql-python-connector==1.4.1
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.0
distlib==0.3.7
docutils==0.20.1
filelock==3.12.2
Flask==3.0.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.43.0
frozenlist==1.4.0
google-auth==2.23.2
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.2
keyring==24.2.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplot==0.1.9
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
mysqlclient==2.2.0
nh3==0.2.14
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pkginfo==1.9.6
platformdirs==3.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyloco==0.0.139
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
SimpleWebSocketServer==0.1.2
six==1.16.0
SQLAlchemy==2.0.21
threadpoolctl==3.2.0
twine==4.0.2
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
ushlex==0.99.1
virtualenv==20.24.3
websocket-client==1.6.3
Werkzeug==3.0.0
yarl==1.9.2
zipp==3.17.0
